chore: rename webcomponents to inapp
