+++ packages/inapp/src/components/mb-button.test.ts
@@ -1,7 +1,7 @@
 import { assert, expect, fixture } from '@open-wc/testing';
 import { html } from 'lit/static-html.js';
 
-import Button from '../dist/mb-button.js';
+import Button from '../../dist/mb-button.js';
 
 suite('mb-button', () => {
   test('is defined', () => {
